Move SnmpTranslate to LibreNMS/Util
