+++ LibreNMS/Data/Source/NetSnmpQuery.php
@@ -28,12 +28,18 @@ namespace LibreNMS\Data\Source;
 
 use App\Facades\LibrenmsConfig;
 use App\Models\Device;
+use App\Models\Eventlog;
+use App\Polling\Measure\Measurement;
 use DeviceCache;
 use Illuminate\Support\Arr;
+use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Str;
+use LibreNMS\Enum\Severity;
+use LibreNMS\Util\Debug;
 use LibreNMS\Util\Oid;
 use LibreNMS\Util\Rewrite;
 use Log;
+use Symfony\Component\Process\Process;
 
 class NetSnmpQuery implements SnmpQueryInterface
 {
